+++ app/models/concerns/thredded/topic_common.rb
@@ -56,6 +56,10 @@ module Thredded
           .merge(reads_class.where(reads[:id].eq(nil).or(reads[:read_at].lt(topics[:last_post_at]))))
       end
 
+      def post_class
+        reflect_on_association(:posts).klass
+      end
+
       private
 
       # @param user [Thredded.user_class]
@@ -63,6 +67,7 @@ module Thredded
       def read_states_by_postable_hash(user)
         read_states = reflect_on_association(:user_read_states).klass
           .where(user_id: user.id, postable_id: current_scope.map(&:id))
+          .include_first_unread(posts_scope: Pundit.policy_scope(user, post_class.all))
         Thredded::TopicCommon::CachingHash.from_relation(read_states)
       end