+++ app/models/thredded/null_user_topic_read_state.rb
@@ -13,5 +13,13 @@ module Thredded
     def post_read?(_post)
       false
     end
+
+    def first_unread_post_id
+      nil
+    end
+
+    def first_unread_post_page
+      nil
+    end
   end
 end